fixed x86_64 compilation
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@927 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/packet_base.cpp
@@ -480,7 +480,7 @@ Certificate PacketBase::ReadCertificate()
 
 	try
 	{
-		cert.LoadRaw(str, str_size);
+		cert.LoadRaw(str, (size_t)str_size);
 	}
 	catch(Certificate::BadCertificate &e)
 	{
